+++ src/main/java/org/example/steps/StoreServiceSteps.java
@@ -10,7 +10,7 @@ import static org.example.service.uritemplate.StoreServiceUri.ORDER_BY_ID;
 public class StoreServiceSteps {
     private static final StoreService STORE_SERVICE = StoreService.getInstance();
 
-    public static Response getOrderById(int orderId) {
+    public static Response getOrderById(long orderId) {
         return STORE_SERVICE.getRequest(ORDER_BY_ID, orderId);
     }
 
@@ -18,7 +18,7 @@ public class StoreServiceSteps {
         return STORE_SERVICE.postRequest(ORDER, order);
     }
 
-    public static void deleteOrderById(int orderId) {
+    public static void deleteOrderById(long orderId) {
         STORE_SERVICE.deleteRequest(ORDER_BY_ID, orderId);
     }
 }
